Add OLTP clients

+++ src/include/storage/proc.h
@@ -115,7 +115,7 @@ struct PGPROC
 								 * else InvalidLocalTransactionId */
 	int			pid;			/* Backend's process ID; 0 if prepared xact */
 	int			pgprocno;
-    uint8_t     rank;           /* Wait priority */
+    double      rank;           /* Wait priority */
 
 	/* These fields are zero while a backend is still starting up: */
 	BackendId	backendId;		/* This backend's backend ID (if assigned) */
